Changed every type

+++ src/gfx/renderer.h
@@ -14,13 +14,13 @@ struct Renderer
     Camera camera;
     Shader shader;
     Chunk* chunks;
-    int chunk_count;
+    u16 chunk_count;
 };
 
 typedef struct Renderer Renderer;
 
 void renderer_init(vec2i viewport_size);
-void renderer_camera_move(vec3f moving, float dt);
+void renderer_camera_move(vec3f moving, f32 dt);
 void renderer_camera_turn(vec2f offset);
 void renderer_render();
 void renderer_reload();
