+++ src/gfx/vao.c
@@ -12,7 +12,7 @@ void vao_bind(VAO vao)
     glBindVertexArray(vao.ID);
 }
 
-void vao_attr(int index, int count, int size, void* offset)
+void vao_attr(u8 index, u8 count, u32 size, void* offset)
 {
     glVertexAttribPointer(index, count, GL_FLOAT, GL_FALSE, size, offset);
     glEnableVertexAttribArray(index);
